Ensure Aftersave receives valid object
This commit is a follow-up to #65 (Commit b42ed240d10a25),
which added support for Aftersave hooks.

There was a flaw in the implementation where the saved
Parse object that was passed into the Aftersave hook
did not have the `createdAt`, `updatedAt`, and `id`
properties.

The solution was to pass the response object to the after-
-save hook, as the missing props were only assigned to
that object.

This commit also resolves issue #67
(https://github.com/HustleInc/parse-mockdb/issues/67)
as the `isNew()` method relies on the presence of the
`id` property.