Remove dependency on `parse-shim`
The usage of `parse-shim` here is a disingenuous; we declare a
peer-dependency on `parse`, but a dependency on `parse-shim`, which
itself has a dependency on `parse`. This can lead to the embedding
application using a different instance of the `parse` module than MockDB
is requiring, with the Parse SDK embedded at different levels of the
node_modules tree.

Removing the parse-shim dependency also means that an application that
is purely node or purely browser based can drop that dependency and just
require the appropriate entry point (folks trying to build an isomorphic
javascript app can still use `parse-shim`).

This does put a little extra burden on the embedding application: they
must explicitly pass a reference to the Parse module into a call to
MockDB.